ints -> longs
git-svn-id: http://svn.vuze.com/public/client@4095 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/tracker/server/impl/TRTrackerServerImpl.java
@@ -123,9 +123,9 @@ TRTrackerServerImpl
 	
 	protected Map		torrent_map = new HashMap(); 
 	
-	protected int		current_announce_retry_interval;
-	protected int		current_scrape_retry_interval;
-	protected int		current_total_clients;
+	protected long		current_announce_retry_interval;
+	protected long		current_scrape_retry_interval;
+	protected long		current_total_clients;
 	
 	protected int		current_min_poll_interval;
